requirements.txt
python-dotenv==1.0.0
openai==1.12.0
httpx[socks]==0.27.2
